feat: refactor v2 -> root, rename evaluate, generate
