+++ libs/ape-common/ape/common/__init__.py
@@ -1,7 +1,7 @@
 from .prompt.cost_tracker import CostTracker, CostTrackerContext
 from .prompt import Prompt
-from .evaluate import Evaluate
-from .generate import BaseGenerate, Generate
+from .evaluator import Evaluator
+from .generator import BaseGenerator, Generator
 from .metric import BaseMetric
 from .global_metric import BaseGlobalMetric
 from .types import MetricResult, GlobalMetricResult, DatasetItem
@@ -11,9 +11,9 @@ from .types.dataset_item import DatasetItem
 __all__ = [
     "CostTracker",
     "Prompt",
-    "Evaluate",
-    "Generate",
-    "BaseGenerate",
+    "Evaluator",
+    "Generator",
+    "BaseGenerator",
     "BaseMetric",
     "BaseGlobalMetric",
     "MetricResult",
