--- libs/ape-common/ape/common/evaluate/__init__.py
@@ -1,3 +1,0 @@
-from .evaluate import Evaluate
-
-__all__ = ["Evaluate"]
