+++ libs/ape-common/ape/common/generator/generator.py
@@ -1,11 +1,11 @@
 from typing import Any, Dict, Type, Union
 
 from pydantic import BaseModel
-from ape.common.generate.generate_base import BaseGenerate
+from ape.common.generator.base_generator import BaseGenerator
 from ape.common.prompt.prompt_base import Prompt
 
 
-class Generate(BaseGenerate):
+class Generator(BaseGenerator):
     async def generate(
         self, prompt: Prompt, inputs: Dict[str, Any] = {}
     ) -> Union[str, Dict[str, Any], Type[BaseModel]]:
