--- libs/ape-core/ape/core/__init__.py
@@ -1,10 +1,0 @@
-from .optimizer.mipro.mipro import MIPRO
-from .optimizer.mipro.mipro_with_hil import MIPROWithHIL
-from .optimizer.mipro.mipro_instruct import MIPROInstruct
-
-
-__all__ = [
-    "MIPRO",
-    "MIPROWithHIL",
-    "MIPROInstruct",
-]
