--- libs/ape-core/ape/core/optimizer/__init__.py
@@ -1,3 +1,0 @@
-from .storage import OptunaSingletonStorage
-
-__all__ = ["OptunaSingletonStorage"]
