--- libs/ape-core/ape/core/proposer/propose_base.py
@@ -1,13 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class Proposer(ABC):
-    def __init__(self):
-        pass
-
-    @abstractmethod
-    async def propose_prompts(self):
-        pass
-
-    async def propose_one(self):
-        pass
