+++ libs/ape-core/ape/core/trainer/fewshot_trainer.py
@@ -1,23 +1,23 @@
 import asyncio
 import copy
 import random
+from typing import Any, List, Optional, Tuple
+
 import numpy as np
-from typing import Any, Dict, List, Literal, Optional, Tuple
-from ape.common.prompt.prompt_base import Prompt
-from ape.common.types import GlobalMetricResult, MetricResult, DatasetItem
-from ape.common.generate import BaseGenerate
+
+from ape.common.generator import BaseGenerator
 from ape.common.global_metric import BaseGlobalMetric
 from ape.common.metric import BaseMetric
-from ape.core.core_prompts import ApeCorePrompts
-from ape.core.proposer.utils import extract_prompt
-from ape.core.v2.trainer.base import BaseTrainer
-from ape.core.v2.types.report import FewShotTrainerReport
+from ape.common.prompt import Prompt
+from ape.common.types import MetricResult, DatasetItem
+from ape.core.trainer.base import BaseTrainer
+from ape.core.types.report import FewShotTrainerReport
 
 
 class FewShotTrainer(BaseTrainer):
     def __init__(
         self,
-        generator: BaseGenerate,
+        generator: BaseGenerator,
         metric: BaseMetric,
         global_metric: BaseGlobalMetric,
         max_bootstrapped_demos: int = 5,
