+++ libs/ape-core/pyproject.toml
@@ -18,7 +18,7 @@ classifiers = [
     "Operating System :: OS Independent",
 ]
 dependencies = [
-    "ape-common>=0.1.3,<0.2.0",
+    "ape-common>=0.1.3",
     "optuna>=4.0.0,<5.0.0",
     "psycopg2-binary>=2.9.9,<3.0.0",
     "sqlalchemy>=2.0.35,<3.0.0",
